+++ src/v1/BasicTypes/Any/mustBeAny.ts
@@ -31,6 +31,27 @@
 // ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 //
+import { DataGuaranteeOptions, TypeGuarantee } from "../../Archetypes";
+import { THROW_THE_ERROR } from "../../ErrorHandling";
+import { DEFAULT_DATA_PATH } from "../../SupportingTypes";
 
-export * from "./isAny";
-export * from "./validateAny";
+/**
+ * `mustBeAny()` is a {@link TypeGuarantee}. Use it to accept any kind
+ * of data whatsoever.
+ *
+ * @param input
+ * the value to inspect
+ * @param onError
+ * Validation never fails, so we never actually call this.
+ * @returns
+ * - `input`
+ *
+ * @category BasicTypes
+ */
+export const mustBeAny: TypeGuarantee<any, DataGuaranteeOptions> = (
+    input: unknown,
+    {
+        onError = THROW_THE_ERROR,
+        path = DEFAULT_DATA_PATH,
+    }: Partial<DataGuaranteeOptions> = {}
+) => input;