+++ CHANGELOG.md
@@ -24,7 +24,8 @@ The following changes have been completed, and will be included in the next tagg
 ### New
 
 * Basic Types
-  - Added `validateAny()'
+  - added `mustBeAny()`
+  - added `validateAny()`
 
 ## v0.1.12