+++ CHANGELOG.md
@@ -21,6 +21,16 @@ For each release, changes are grouped under these headings:
 
 The following changes have been completed, and will be included in the next tagged release.
 
+## v0.1.13
+
+Released Wednesday, 8th July 2020.
+
+### New
+
+* Basic Types
+  - added `mustBeAny()`
+  - added `validateAny()`
+
 ## v0.1.12
 
 Released Wednesday, 1st July 2020.